requirements.txt
aioice==0.7.6
aiortc==1.3.2
altair==4.2.0
attrs==22.1.0
av==9.2.0
blinker==1.5
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
cryptography==38.0.4
decorator==5.1.1
dnspython==2.2.1
entrypoints==0.4
ffmpeg-python==0.2.0
filelock==3.8.2
future==0.18.2
gitdb==4.0.10
GitPython==3.1.29
google-crc32c==1.5.0
huggingface-hub==0.11.1
idna==3.4
importlib-metadata==5.1.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.1
more-itertools==9.0.0
netifaces==0.11.0
numpy==1.23.0
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
protobuf==3.20.3
pyarrow==10.0.1
pycparser==2.21
pydeck==0.8.0
pydub==0.25.1
pyee==9.0.4
Pygments==2.13.0
pylibsrtp==0.7.1
Pympler==1.0.1
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
rich==12.6.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.10.0
streamlit-webrtc==0.43.0
tokenizers==0.13.2
toml==0.10.2
toolz==0.12.0
torch==1.13.0
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
transformers==4.25.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
watchdog==2.2.0
git+https://github.com/openai/whisper.git@main#egg=whisper
zipp==3.11.0
